refactor: move implementations to modules
