+++ crates/rattler_solve/src/libsolv_sys/wrapper/flags.rs
