+++ crates/rattler_solve/src/libsolv_sys/wrapper/mod.rs
