+++ crates/rattler_solve/src/libsolv_sys/wrapper/solve_problem.rs
